requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0
django-cors-headers==4.4.0
